Create projIdea
In this  Step, we will build a custom Docker image which contains a simple FLASK based web application. The custom Docker image will later be deployed into the Kubernetes cluster. Before performing the deployment within the cluster, 'll be shown how to create a new namespace resource and how to set it as the default namespace into which the remaining deployment takes place.

Apply Nginx ConfigMap
In this  Step, 'll learn how to update and edit a manifest file to create a new ConfigMap resource within the cluster. ConfigMaps allow you to decouple configuration artifacts from image content to keep containerized applications portable. In this scenario a ConfigMap will be used to store the NGINX web server configuration externally from the NGINX container image - allowing it to be passed in at launch time. 

apply deployment 
In this  Step, y'll learn how to update and edit a manifest file to create a new Deployment resource within the cluster. Deployments represent a set of multiple, identical Pods and have the ability to automatically replace failing or unresponsive instances. In this scenario a Deployment will be used to launch a single pod containing two containers. The first container is a NGINX web server container which proxies incoming HTTP requests downstream to a second container running a FLASK based web application - which you created in the first  Step.

Create Helm Project

in this  Step, will learn about the basics of Helm and w'll learn how to use the helm create command to scaffold or generate an empty Helm Chart project. You'll then learn to use the helm template command to render a Kubernetes manifest file, which can then be deployed into a Kubernetes cluster.

Notes:

Helm 3 is used
The empty Helm Chart test-app project that you create in this Step will NOT be deployed into the Kubernetes cluster, it is only being used to demonstrate how to create and scaffold new Helm based projects. In the next Step 'll get to deploy an existing Helm templated project into the cluster.

Update Deployment to Use Helm Templates
In this  Step, 'll learn how to use the helm and kubectl commands together to deploy a provided Helm chart into the Kubernetes cluster. In this scenario the provided app Helm chart has been configured to create a Deployment which launches a single Pod containing two containers. The first container is a NGINX web server container which proxies incoming HTTP requests downstream to a second container running a FLASK based web application (based on the Docker image you earlier built). 'll learn how to use a Helm Values File to externalize runtime variables, ensuring that your Helm templates can be reused across different environments, for example say dev and prod environments.

Notes:

Helm 3 is used